src/test/requirements_ai_assistant_fixed.txt
# AI助手依赖包（修复版本）
volcengine-python-sdk[ark]>=1.0.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
opencv-python>=4.8.0
numpy>=1.24.0
pyttsx3>=2.90
pygame>=2.5.0
gtts>=2.3.0
bme680>=1.1.1
pyserial>=3.5
keyboard>=0.13.5
pocketsphinx>=0.1.15